Version bump to 1.5.0, reqProduct 4.3.0
OptionalType is released with product version 4.3.0.
This toolkit version depends on product providing optionalType feature.
It can't be used on older platforms.
Version change only in 2nd place and not in first as optionalType
support is a feature and no major change. All changes to older version
are now bound to 1.4.x and as such are only fixes.